improved stage control candidate
creates a list of active (triggered) timers to track maximum requested
boost levels for each calling thermostat. When a call for heat is
satisfied on a thermostat, a reevaluation of stage level requests is
triggered and the stage level is reset to the maximum level for all
remaining active triggers.